J721E: disable hard-coded reg and irq values
Disabling the hard-coded GPU register base, register size and irq values
as driver should fail if GPU DT node is not present. If it does not
fail, power-domains will fail to start during the initialization and
the driver will cause a kernel panic when it tried to access the device
that does not exist (has not been turned on).

Signed-off-by: Erick Narvaez <a0226667@ti.com>