Virt: HACK: Forward ISR from host to guest
Trigger the interrupt for the guest from the host driver. This makes use
of main_i2c4 device's interrupt. The host will enable the i2c4 device
and then trigger its interrupt whenever it recieves an interrupt from
the GPU. The guest will clear the interrupt as part of its ISR.

Signed-off-by: Erick Narvaez <e-narvaez@ti.com>