Virt: Fix number of PT entries configuration
The mips page table is always setup to have 8192 entries of 4 bytes
mapping a max fw heap size of 32 MB. In the driver, the number of PT
entries are calculated based on the actual fw heap size, which can be
different to 32 MB, causing descrepencies in mapping some allocations.
This was causing crashes in the pvr profiling server tool. Fixing the
calculation of the number of PT entries on the size of page table
instead of actual fw heap size.

IMG#123899

Signed-off-by: Erick Narvaez <e-narvaez@ti.com>