ddk: km: add support of am62_android and android kernel build system[1]
Add Makefile to support Android for am62.

Google provide a kernel build system use to compile modules and kernel.
Link to Generic Kernel Image features we need to use wame env as AOSP.
DDK is not user friendly and it's not possible to use it without adding
specific files.

- add config_kernel.h generate at build time with DDK build sript and
rename it : config_kernel_am62_android.h
- add BUILD.bazel file to support new android kernel build system

[1] : https://source.android.com/setup/build/building-kernels

Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>