bxs_am62p: pull dust power domain into apm
Pull dust domain APM into the driver again. Apparently this appeared to
be working do to some interesting behavior with TI-SCI that I was not
aware of.

The proper solution until a suitable external power controller is
designated for the dust domain is to pull it back into the driver and
create the appropriate links and genpd virtual devices.

Signed-off-by: Randolph Sapp <rs@ti.com>