axe_am62: sysconfig: clock fixup
Fixup clock interactions such that clock rates are set by the driver and
the driver's internal value is updated to reflect this. This internal
value may drift over time as the driver will occasionally poll the
device and attempt to calculate the frequency for it's own internal use.

Use a similar power data structure as the other devices for tracking
clocks and a ref to the device. As this is using a single power domain
genpd takes care of most of the power domain work so we don't need
either of those values for this device.

Signed-off-by: Randolph Sapp <rs@ti.com>