services: server: physmem_dmabuf: fix the ref counter
Rework the reference counter logic for the kernel mapping. A lot of this
seemed to dance around some expectations that weren't always true. This
patch is spiritually equivalent to the below supplied patch but it is
not implemented the same.

Should fix PVRSRV_ERROR_OBJECT_STILL_REFERENCED errors.

Corresponds to:
PP148569_CL6526019.patch

Signed-off-by: Randolph Sapp <rs@ti.com>