cc-check: switch some logic around for GCC 14
Switch some of this logic around so it is compatible with GCC 13 and 14.
GCC 14, when given an invalid program to compile will now return 1 and
not produce an object. On top of that, it has changed the unrecognized
command line flags warning into a note.

Also dump all supported flags to a temporary text file so they can be
checked post compilation.

Signed-off-by: Randolph Sapp <rs@ti.com>