cc-check: add gcc 13 string to grep
Add GCC 13's string for invalid warning options to the grep. GCC 13
states that the check option itself doesn't exist when being asked
to disable an invalid warning.

For example if -Wno-error=implicit-fallthrough is passed and
implicit-fallthrough is not a valid check the returned value says that
-Wimplicit-fallthrough is not a valid option instead of just saying the
given value is invalid.

Signed-off-by: Randolph Sapp <rs@ti.com>